SerializableUtils is now its own file. URISyntaxErrors shouldn't crash anymore.
